encapsulating validations
